backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
celery>=5.4,<6
redis>=5.0,<6
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
bcrypt==4.0.1
python-dotenv==1.0.*
pydantic==2.10.*
langgraph>=0.2.0
langgraph-supervisor>=0.0.31
langgraph-swarm>=0.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
httpx>=0.27.0
lark-oapi>=1.3.0
apscheduler>=3.10.0
cryptography>=42.0.0
RestrictedPython>=7.0
